Add files via upload
A GUI created for my capstone project at Stark State College. The project pulls information entered into a database about the top 10 at each position in the NFL according to statistics. The project was completed over the course of 16 weeks and involved multiple steps not included. Included is the final product of only the code.